Reorganised tests and tweaked getParts
